requirements.txt
# Core scraping
requests>=2.32,<3

# Database (psycopg3)
psycopg[binary]>=3.2,<4

# Config management
pydantic>=2.9,<3
pydantic-settings>=2.6,<3

# Profile files
pyyaml>=6.0,<7

# Notifications — native channels
discord-webhook>=1.3,<2

# Notifications — 100+ channels via a single URL
apprise>=1.9,<2
